split flowchart test file
